+++ php/src/Application/Command/DeduceTypes.php
@@ -86,7 +86,7 @@ class DeduceTypes extends AbstractCommand
         $offset = $arguments['offset']->value;
 
         if (isset($arguments['charoffset']) && $arguments['charoffset']->value == true) {
-            $offset = $this->getSourceCodeHelper()->getCharacterOffsetFromByteOffset($offset, $code);
+            $offset = $this->getSourceCodeHelper()->getByteOffsetFromCharacterOffset($offset, $code);
         }
 
         $result = $this->deduceTypes(